+++ packages/ai-native/src/browser/widget/inline-chat/inline-content-widget.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Autowired, INJECTOR_TOKEN, Injectable, Injector } from '@opensumi/di';
-import { IAIInlineChatService } from '@opensumi/ide-core-browser';
+import { IAIInlineChatService, StackingLevelStr } from '@opensumi/ide-core-browser';
 import { AIInlineChatContentWidget, Emitter } from '@opensumi/ide-core-common';
 import * as monaco from '@opensumi/ide-monaco';
 import { monacoBrowser } from '@opensumi/ide-monaco/lib/browser';
@@ -9,7 +9,6 @@ import {
   ReactInlineContentWidget,
   ShowAIContentOptions,
 } from '@opensumi/ide-monaco/lib/browser/ai-native/BaseInlineContentWidget';
-import { StackingLevelStr } from '@opensumi/ide-theme';
 
 import { AINativeContextKey } from '../../contextkey/ai-native.contextkey.service';
