+++ packages/core-browser/__tests__/window/window.service.test.ts
@@ -1,8 +1,8 @@
 import { WindowService } from '@opensumi/ide-core-browser/lib/window/window.service';
 import { IElectronMainLifeCycleService, IElectronMainUIService } from '@opensumi/ide-core-common/lib/electron';
+import { createBrowserInjector } from '@opensumi/ide-dev-tool/src/injector-helper';
+import { MockInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 
-import { createBrowserInjector } from '../../../../tools/dev-tool/src/injector-helper';
-import { MockInjector } from '../../../../tools/dev-tool/src/mock-injector';
 import { IWindowService, URI } from '../../src';
 import { IExternalUriService } from '../../src/services';
