+++ packages/core-browser/src/design/index.ts
@@ -1,2 +1,3 @@
 export * from './design.style.service';
 export * from './types';
+export * from './rule';
